move 'magic command' code to 'cells'
