+++ tests/unit/source_code/notebooks/test_cells.py
@@ -1,12 +1,15 @@
 from pathlib import Path
 from unittest.mock import create_autospec
+
+import pytest
 import yaml
 
 
 from databricks.labs.ucx.source_code.base import CurrentSessionState
-from databricks.labs.ucx.source_code.graph import Dependency, DependencyGraph, DependencyResolver
+from databricks.labs.ucx.source_code.graph import Dependency, DependencyGraph, DependencyResolver, DependencyProblem
 from databricks.labs.ucx.source_code.linters.files import FileLoader, ImportFileResolver
-from databricks.labs.ucx.source_code.notebooks.cells import CellLanguage, PipCell, PythonCell
+from databricks.labs.ucx.source_code.linters.python_ast import Tree
+from databricks.labs.ucx.source_code.notebooks.cells import CellLanguage, PipCell, PythonCell, PipMagic, MagicCommand
 from databricks.labs.ucx.source_code.notebooks.loaders import (
     NotebookResolver,
     NotebookLoader,
@@ -175,3 +178,45 @@ b = 'else'
     graph = DependencyGraph(dependency, None, simple_dependency_resolver, mock_path_lookup, CurrentSessionState())
     problems = cell.build_dependency_graph(graph)
     assert not problems
+
+
+@pytest.mark.parametrize(
+    "code,split",
+    [
+        ("%pip install foo", ["%pip", "install", "foo"]),
+        ("%pip install", ["%pip", "install"]),
+        ("%pip installl foo", ["%pip", "installl", "foo"]),
+        ("%pip install foo --index-url bar", ["%pip", "install", "foo", "--index-url", "bar"]),
+        ("%pip install foo --index-url bar", ["%pip", "install", "foo", "--index-url", "bar"]),
+        ("%pip install foo --index-url \\\n bar", ["%pip", "install", "foo", "--index-url", "bar"]),
+        ("%pip install foo --index-url bar\nmore code", ["%pip", "install", "foo", "--index-url", "bar"]),
+        (
+            "%pip install foo --index-url bar\\\n -t /tmp/",
+            ["%pip", "install", "foo", "--index-url", "bar", "-t", "/tmp/"],
+        ),
+        ("%pip install foo --index-url \\\n bar", ["%pip", "install", "foo", "--index-url", "bar"]),
+        (
+            "%pip install ./distribution/dist/thingy-0.0.1-py2.py3-none-any.whl",
+            ["%pip", "install", "./distribution/dist/thingy-0.0.1-py2.py3-none-any.whl"],
+        ),
+        (
+            "%pip install distribution/dist/thingy-0.0.1-py2.py3-none-any.whl",
+            ["%pip", "install", "distribution/dist/thingy-0.0.1-py2.py3-none-any.whl"],
+        ),
+    ],
+)
+def test_pip_command_split(code, split):
+    assert PipMagic._split(code) == split  # pylint: disable=protected-access
+
+
+def test_unsupported_magic_raises_problem(simple_dependency_resolver, mock_path_lookup):
+    source = """
+%unsupported stuff '"%#@!
+"""
+    converted = Tree.convert_magic_lines_to_magic_commands(source)
+    tree = Tree.parse(converted)
+    commands, _ = MagicCommand.extract_from_tree(tree, DependencyProblem.from_node)
+    dependency = Dependency(FileLoader(), Path(""))
+    graph = DependencyGraph(dependency, None, simple_dependency_resolver, mock_path_lookup, CurrentSessionState())
+    problems = commands[0].build_dependency_graph(graph)
+    assert problems[0].code == "unsupported-magic-line"
